Taking advantage of the talon follow mode
PID drive now uses 2 PIDControllers (right now 1 because we have a
missing encoder in the robot) insted of 4, thanks to the option of talon
follow mode in "talonControlMode".
This mode will make a talon follow a specific talon (see lines 78-84 in
drive train).
PID vision uses 1 PIDController insted  of 4 for the same reason (see
lines 60-68 in vision).

All PID related static member are now static only and not final because
I had to add an option to change them in the SmartDeshboard in run time.

Some clean up has been made.

**CODE NOT TESTED**
Code is finally ready to be tested! :sparkles: :star2: