Added Feed-Forward
Added feed-forward to the PID calculation. This means that now the talon
has a minimum ouput sent to it, which makes any additional output move
the motor, which makes the PID way more sensitive, As a result, the PID
was tuned down, and is now way faster and more responsive.

**Shooting Wheel is canceled due to mechanical problems, wait for fix to
turn it back on**