Out put range of vision has been down scaled
The out put range of the vision PID changed to a softer one so it will
be easier for the robot to center on the target when the camera does not
see the target.
D value has been chaged so it will be softer.

ShootByVision will only call to CenterByVision until the shooter wheel
will be connected.